create tools.fluxdiagram module and move stuff from the scripts module
refactor the arg parsing into its own function.
